requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0

# APIs
tweepy==4.14.0
google-generativeai
groq

# Security & Config
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Retry Logic
tenacity==8.2.3

# Database
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Testing (Optional - uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
httpx==0.27.0

# Logging

# Authentication
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
